pyproject.toml
[project]
name = "repl-nix-datenightmagic"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "openai>=1.52.2",
    "torch>=2.5.1",
    "transformers>=4.46.1",
    "google-cloud-vision>=3.8.0",
    "pillow>=11.0.0",
    "mediapipe>=0.10.15",
    "numpy>=1.26.4",
    "opencv-python>=4.10.0.84",
    "opencv-python-headless",
    "werkzeug>=3.0.6",
    "requests",
    "flask-session>=0.8.0",
    "sqlalchemy",
    "xds-protos",
    "psutil>=6.1.0",
]
